Initial Commit ( Version 1.2.0 )
Version 1.2.0 of a little sideproject of mine called MEMEPAINT.
Basically it draws images of memes and somtimes it plays some audio, it
is written in pure clientside code ( HTML, CSS, and JavaScript ). Have
fun with it! You can find a live version at memepaint.firebaseapp.com or
for a development version go to memepaint-pietroglyph.c9users.io.